backend/pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.121.2",
    "httpx>=0.28.0",
    "jinja2>=3.1.6",
    "logfire[fastapi]>=4.15.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "scalar-fastapi>=1.0.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.121.2",
    "locust>=2.42.0",
    "pre-commit>=4.5.0",
    "psycopg[binary]>=3.3.2",
    "pytest>=9.0.1",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
]
